requirements.txt
streamlit
langchain
langchain_groq
langchain-community
python-dotenv
replicate